creating and using rpc mint and burn everywhere
